Switching back to the old style of DDS messages.
The new style of building them in projects wasn't really compatible with partial builds, because we often need to use messages defined in a project that we don't actually want to
build. Allowing projects to be partially built, supplying only their DDS messages would make the top level build logic fairly complex and convoluted. Centralizing all idls into
one project also makes it easy to generate a library for python, though this could've been done in the old approach, albeit with more effort and complexity.

The main downside to doing this is build efficiency. Unfortunately, we're back to the situation where touching a single DDS message relinks every single executable again.
Hopefully this won't be happening as often as it was last year now that we've had a lot more time to think and plan out workers, since getting rid of this problem doesn't seem to
be
worth the complexity it brings on.