Revert "made py_modules add the build lib directories to LD_LIBRARY_PATH so that scripts can find DDS."
It messes up the script on my system, and isn't really in the initial intent of the script. If you need it to make your
setup work .gitignore it or use an alias or something, but all this script is supposed to do is allow python modules in
the source directory to be found. Changes to C++ code still requires a make install, which takes almost exactly the same
amount of time as running make.

This reverts commit 10dfbfdb150bb91083947ac3e7d7a4f6906e1a40.